Added the beginnings of a new level editor. A little more cleanup. Updated the MSVC 2003 project file that was in SVN (it was very out of date). Someone should double-check my Makefile.am changes because I'm developing with MSVC, and not on Linux.  It looks right, but I'm unsure.

+++ src/display/drawing_context.cxx
@@ -63,7 +63,7 @@ public:
   virtual ~SurfaceDrawingRequest() {}
 
   void draw(CL_GraphicContext* gc) {
-    sprite.draw(static_cast<int>(pos.x), static_cast<int>(pos.y), gc);
+    sprite.draw(pos.x, pos.y, gc);
   }
 };
 
@@ -150,7 +150,7 @@ public:
   virtual ~SpriteDrawingRequest() {}
 
   void draw(CL_GraphicContext* gc) {
-    sprite.draw(static_cast<int>(pos.x), static_cast<int>(pos.y), gc);
+    sprite.draw(pos.x, pos.y, gc);
   }
 };
